Removing unused code (#1421)
