+++ src/selectors/getLayoutData.ts
@@ -5,8 +5,6 @@ import type { RootState } from 'src/redux/store';
 
 const selectFocusedComponent = (state: RootState) => state.formLayout.uiConfig.focus;
 
-const selectHiddenFields = (state: RootState) => state.formLayout.uiConfig.hiddenFields;
-
 const selectId = (_state, props) => props.id;
 
 export const makeGetFocus = (): ParametricSelector<
@@ -16,11 +14,3 @@ export const makeGetFocus = (): ParametricSelector<
   },
   boolean
 > => createSelector([selectFocusedComponent, selectId], (focus, id) => focus === id);
-
-export const makeGetHidden = (): ParametricSelector<
-  RootState,
-  {
-    id: string;
-  },
-  boolean
-> => createSelector([selectHiddenFields, selectId], (fields, id) => fields.findIndex((itemId) => itemId === id) > -1);
