+++ src/selectors/simpleSelectors.test.ts
@@ -1,20 +1,14 @@
 import { getFormLayoutStateMock } from 'src/__mocks__/formLayoutStateMock';
 import { getInitialStateMock } from 'src/__mocks__/initialStateMock';
-import { getProcessStateMock } from 'src/__mocks__/processMock';
 import {
   appMetaDataSelector,
   currentSelectedPartyIdSelector,
   instanceDataSelector,
   layoutSetsSelector,
-  processStateSelector,
-  profileStateSelector,
 } from 'src/selectors/simpleSelectors';
 
 describe('simpleSelectors', () => {
   const state = getInitialStateMock({
-    process: getProcessStateMock({
-      taskId: 'Task1',
-    }),
     formLayout: {
       ...getFormLayoutStateMock(),
       layoutsets: {
@@ -38,11 +32,6 @@ describe('simpleSelectors', () => {
     expect(instance?.id).toBe('91cefc5e-c47b-40ff-a8a4-05971205f783');
   });
 
-  it('should fetch process state', () => {
-    const process = processStateSelector(state);
-    expect(process.taskId).toBe('Task1');
-  });
-
   it('should fetch selected party id', () => {
     const selectedPartyId = currentSelectedPartyIdSelector(state);
     expect(selectedPartyId).toBe('12345');
@@ -52,10 +41,5 @@ describe('simpleSelectors', () => {
     const layoutSets = layoutSetsSelector(state);
     expect(layoutSets?.sets.length).toBe(1);
     expect(layoutSets?.sets[0].id).toBe('layout');
-  });
-
-  it('should fetch profile', () => {
-    const profile = profileStateSelector(state);
-    expect(profile.partyId).toBe(12345);
   });
 });
